refactor: change the file structure a bit
