+++ src/handlers/index.ts
@@ -3,9 +3,8 @@
  * See COPYING.txt for license details.
  */
 
-export { default as addToCartHandler } from "./addToCart";
-export { default as placeOrderHandler } from "./placeOrder";
-export { default as productViewHandler } from "./productView";
+export * from "./checkout";
+export * from "./product";
 export * from "./recommendations";
 export * from "./search";
-export { default as shoppingCartViewHandler } from "./shoppingCartView";
+export * from "./shoppingCart";
