+++ src/handlers/product/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+
+export { default as addToCartHandler } from "./addToCart";
+export { default as productViewHandler } from "./view";
