+++ src/handlers/shoppingCart/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+
+export { default as shoppingCartViewHandler } from "./view";
