+++ src/handlers/shoppingCart/view.ts
@@ -6,23 +6,20 @@
 import { Event } from "@adobe/magento-storefront-events-sdk/dist/types/types/events";
 import { trackStructEvent } from "@snowplow/browser-tracker";
 
-import { createProductCtx, createShoppingCartCtx } from "../contexts";
+import { createShoppingCartCtx } from "../../contexts";
 
 const handler = (event: Event): void => {
-    const {
-        pageContext,
-        productContext,
-        shoppingCartContext,
-    } = event.eventInfo;
-
-    const productCtx = createProductCtx(productContext);
+    const { pageContext, shoppingCartContext } = event.eventInfo;
     const shoppingCartCtx = createShoppingCartCtx(shoppingCartContext);
 
     trackStructEvent({
-        category: "product",
+        category: "shopping-cart",
         action: "view",
         property: pageContext.pageType,
-        context: [productCtx, shoppingCartCtx],
+        // TODO: this should be the cartId, which is a string,
+        //       but Snowplow expects a number for value.
+        value: 0,
+        context: [shoppingCartCtx],
     });
 };
